+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG for Sulu
     * BUGFIX      #4540  [ContentBundle]        Fix copying of teaser selection in WebspaceCopyCommand
     * FEATURE     #4532  [MediaBundle]          Add command to remove format cache files which not longer exists
     * HOTFIX      #4512  [Components]           Fix pathcleanup whitespace character problems
+    * FEATURE     #4504  [WebsiteBundle]        Dispatch an event after cache clear. 
 
 * 1.6.26 (2019-03-26)
     * BUGFIX      #4343  [MediaBundle]          Fixed format manager types configuration cannot be changed